move header files into header component folder
